@@ -9,9 +9,9 @@
     "config": {
         "sort-packages": true
     },
-    "version": "100.3.6",
+    "version": "100.3.7",
     "require": {
-        "php": "~7.1.3||~7.2.0||~7.3.0",
+        "php": "~7.3.0||~7.4.0",
         "magento/framework": "102.0.*",
         "magento/module-backend": "101.0.*"
     },
