@@ -22,6 +22,8 @@ use Symfony\Component\Console\Tester\CommandTester;
  */
 class IndexerReindexCommandTest extends AbstractIndexerCommandCommonSetup
 {
+    const STUB_INDEXER_NAME = 'Indexer Name';
+
     /**
      * Command being tested
      *
@@ -30,17 +32,17 @@ class IndexerReindexCommandTest extends AbstractIndexerCommandCommonSetup
     private $command;
 
     /**
-     * @var \Magento\Framework\Indexer\ConfigInterface|\PHPUnit_Framework_MockObject_MockObject
+     * @var \Magento\Framework\Indexer\ConfigInterface|\PHPUnit\Framework\MockObject\MockObject
      */
     protected $configMock;
 
     /**
-     * @var IndexerRegistry|\PHPUnit_Framework_MockObject_MockObject
+     * @var IndexerRegistry|\PHPUnit\Framework\MockObject\MockObject
      */
     private $indexerRegistryMock;
 
     /**
-     * @var DependencyInfoProvider|\PHPUnit_Framework_MockObject_MockObject
+     * @var DependencyInfoProvider|\PHPUnit\Framework\MockObject\MockObject
      */
     private $dependencyInfoProviderMock;
 
@@ -52,7 +54,7 @@ class IndexerReindexCommandTest extends AbstractIndexerCommandCommonSetup
     /**
      * Set up
      */
-    public function setUp()
+    protected function setUp(): void
     {
         $this->objectManagerHelper = new ObjectManagerHelper($this);
         $this->configMock = $this->createMock(\Magento\Indexer\Model\Config::class);
@@ -86,7 +88,7 @@ class IndexerReindexCommandTest extends AbstractIndexerCommandCommonSetup
         $this->stateMock->expects($this->never())->method('setAreaCode');
         $this->command = new IndexerReindexCommand($this->objectManagerFactory);
         $optionsList = $this->command->getInputList();
-        $this->assertSame(1, count($optionsList));
+        $this->assertCount(1, $optionsList);
         $this->assertSame('index', $optionsList[0]->getName());
     }
 
@@ -94,13 +96,11 @@ class IndexerReindexCommandTest extends AbstractIndexerCommandCommonSetup
     {
         $this->configMock->expects($this->once())
             ->method('getIndexer')
-            ->will(
-                $this->returnValue(
-                    [
-                        'title' => 'Title_indexerOne',
-                        'shared_index' => null
-                    ]
-                )
+            ->willReturn(
+                [
+                    'title' => 'Title_indexerOne',
+                    'shared_index' => null
+                ]
             );
         $this->configureAdminArea();
         $this->initIndexerCollectionByItems(
@@ -187,7 +187,7 @@ class IndexerReindexCommandTest extends AbstractIndexerCommandCommonSetup
             );
         }
         $pattern .= '$#';
-        $this->assertRegExp($pattern, $commandTester->getDisplay());
+        $this->assertMatchesRegularExpression($pattern, $commandTester->getDisplay());
     }
 
     /**
@@ -221,11 +221,11 @@ class IndexerReindexCommandTest extends AbstractIndexerCommandCommonSetup
      * @param array|null $methods
      * @param array $data
      *
-     * @return \PHPUnit_Framework_MockObject_MockObject|StateInterface
+     * @return \PHPUnit\Framework\MockObject\MockObject|StateInterface
      */
     private function getStateMock(array $methods = null, array $data = [])
     {
-        /** @var \PHPUnit_Framework_MockObject_MockObject|StateInterface $state */
+        /** @var \PHPUnit\Framework\MockObject\MockObject|StateInterface $state */
         $state = $this->getMockBuilder(StateInterface::class)
             ->setMethods($methods)
             ->disableOriginalConstructor()
@@ -405,9 +405,12 @@ class IndexerReindexCommandTest extends AbstractIndexerCommandCommonSetup
     public function testExecuteWithLocalizedException()
     {
         $this->configureAdminArea();
-        $indexerOne = $this->getIndexerMock(['reindexAll', 'getStatus'], ['indexer_id' => 'indexer_1']);
+        $indexerOne = $this->getIndexerMock(
+            ['reindexAll', 'getStatus'],
+            ['indexer_id' => 'indexer_1', 'title' => self::STUB_INDEXER_NAME]
+        );
         $localizedException = new LocalizedException(new Phrase('Some Exception Message'));
-        $indexerOne->expects($this->once())->method('reindexAll')->will($this->throwException($localizedException));
+        $indexerOne->expects($this->once())->method('reindexAll')->willThrowException($localizedException);
         $this->initIndexerCollectionByItems([$indexerOne]);
         $this->command = new IndexerReindexCommand($this->objectManagerFactory);
         $commandTester = new CommandTester($this->command);
