@@ -13,17 +13,17 @@ class DataTest extends \PHPUnit\Framework\TestCase
     protected $model;
 
     /**
-     * @var \Magento\Framework\Indexer\Config\Reader|\PHPUnit_Framework_MockObject_MockObject
+     * @var \Magento\Framework\Indexer\Config\Reader|\PHPUnit\Framework\MockObject\MockObject
      */
     protected $reader;
 
     /**
-     * @var \Magento\Framework\Config\CacheInterface|\PHPUnit_Framework_MockObject_MockObject
+     * @var \Magento\Framework\Config\CacheInterface|\PHPUnit\Framework\MockObject\MockObject
      */
     protected $cache;
 
     /**
-     * @var \Magento\Indexer\Model\ResourceModel\Indexer\State\Collection|\PHPUnit_Framework_MockObject_MockObject
+     * @var \Magento\Indexer\Model\ResourceModel\Indexer\State\Collection|\PHPUnit\Framework\MockObject\MockObject
      */
     protected $stateCollection;
 
@@ -38,11 +38,11 @@ class DataTest extends \PHPUnit\Framework\TestCase
     protected $indexers = ['indexer1' => [], 'indexer3' => []];
 
     /**
-     * @var \Magento\Framework\Serialize\SerializerInterface|\PHPUnit_Framework_MockObject_MockObject
+     * @var \Magento\Framework\Serialize\SerializerInterface|\PHPUnit\Framework\MockObject\MockObject
      */
     private $serializerMock;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->reader = $this->createPartialMock(\Magento\Framework\Indexer\Config\Reader::class, ['read']);
         $this->cache = $this->getMockForAbstractClass(
@@ -64,7 +64,7 @@ class DataTest extends \PHPUnit\Framework\TestCase
     public function testConstructorWithCache()
     {
         $serializedData = 'serialized data';
-        $this->cache->expects($this->once())->method('test')->with($this->cacheId)->will($this->returnValue(true));
+        $this->cache->expects($this->once())->method('test')->with($this->cacheId)->willReturn(true);
         $this->cache->expects($this->once())
             ->method('load')
             ->with($this->cacheId)
@@ -88,28 +88,28 @@ class DataTest extends \PHPUnit\Framework\TestCase
 
     public function testConstructorWithoutCache()
     {
-        $this->cache->expects($this->once())->method('test')->with($this->cacheId)->will($this->returnValue(false));
-        $this->cache->expects($this->once())->method('load')->with($this->cacheId)->will($this->returnValue(false));
+        $this->cache->expects($this->once())->method('test')->with($this->cacheId)->willReturn(false);
+        $this->cache->expects($this->once())->method('load')->with($this->cacheId)->willReturn(false);
 
-        $this->reader->expects($this->once())->method('read')->will($this->returnValue($this->indexers));
+        $this->reader->expects($this->once())->method('read')->willReturn($this->indexers);
 
         $stateExistent = $this->createPartialMock(
             \Magento\Indexer\Model\Indexer\State::class,
             ['getIndexerId', '__wakeup', 'delete']
         );
-        $stateExistent->expects($this->once())->method('getIndexerId')->will($this->returnValue('indexer1'));
+        $stateExistent->expects($this->once())->method('getIndexerId')->willReturn('indexer1');
         $stateExistent->expects($this->never())->method('delete');
 
         $stateNonexistent = $this->createPartialMock(
             \Magento\Indexer\Model\Indexer\State::class,
             ['getIndexerId', '__wakeup', 'delete']
         );
-        $stateNonexistent->expects($this->once())->method('getIndexerId')->will($this->returnValue('indexer2'));
+        $stateNonexistent->expects($this->once())->method('getIndexerId')->willReturn('indexer2');
         $stateNonexistent->expects($this->once())->method('delete');
 
         $states = [$stateExistent, $stateNonexistent];
 
-        $this->stateCollection->expects($this->once())->method('getItems')->will($this->returnValue($states));
+        $this->stateCollection->expects($this->once())->method('getItems')->willReturn($states);
 
         $this->model = new \Magento\Indexer\Model\Config\Data(
             $this->reader,
