@@ -15,11 +15,11 @@ class AbstractProcessorTest extends \PHPUnit\Framework\TestCase
     protected $model;
 
     /**
-     * @var \Magento\Framework\Indexer\IndexerRegistry|\PHPUnit_Framework_MockObject_MockObject
+     * @var \Magento\Framework\Indexer\IndexerRegistry|\PHPUnit\Framework\MockObject\MockObject
      */
     protected $_indexerRegistryMock;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->_indexerRegistryMock = $this->createPartialMock(
             \Magento\Framework\Indexer\IndexerRegistry::class,
@@ -73,14 +73,14 @@ class AbstractProcessorTest extends \PHPUnit\Framework\TestCase
                 self::INDEXER_ID
             )->willReturnSelf();
             $this->_indexerRegistryMock->expects($this->once())->method('isScheduled')->willReturn($scheduled);
-            $this->assertEquals(null, $this->model->reindexRow($id));
+            $this->assertNull($this->model->reindexRow($id));
         } else {
             $this->_indexerRegistryMock->expects($this->exactly(2))->method('get')->with(
                 self::INDEXER_ID
             )->willReturnSelf();
             $this->_indexerRegistryMock->expects($this->once())->method('isScheduled')->willReturn($scheduled);
             $this->_indexerRegistryMock->expects($this->once())->method('reindexRow')->with($id)->willReturnSelf();
-            $this->assertEquals(null, $this->model->reindexRow($id));
+            $this->assertNull($this->model->reindexRow($id));
         }
     }
 
@@ -96,14 +96,14 @@ class AbstractProcessorTest extends \PHPUnit\Framework\TestCase
                 self::INDEXER_ID
             )->willReturnSelf();
             $this->_indexerRegistryMock->expects($this->once())->method('isScheduled')->willReturn($scheduled);
-            $this->assertEquals(null, $this->model->reindexList($ids));
+            $this->assertNull($this->model->reindexList($ids));
         } else {
             $this->_indexerRegistryMock->expects($this->exactly(2))->method('get')->with(
                 self::INDEXER_ID
             )->willReturnSelf();
             $this->_indexerRegistryMock->expects($this->once())->method('isScheduled')->willReturn($scheduled);
             $this->_indexerRegistryMock->expects($this->once())->method('reindexList')->with($ids)->willReturnSelf();
-            $this->assertEquals(null, $this->model->reindexList($ids));
+            $this->assertNull($this->model->reindexList($ids));
         }
     }
 
