@@ -21,12 +21,12 @@ class CollectionTest extends \PHPUnit\Framework\TestCase
         $connectionMock = $this->createMock(\Magento\Framework\DB\Adapter\Pdo\Mysql::class);
         $selectRendererMock = $this->createMock(\Magento\Framework\DB\Select\SelectRenderer::class);
         $resourceMock = $this->createMock(\Magento\Framework\Flag\FlagResource::class);
-        $resourceMock->expects($this->any())->method('getConnection')->will($this->returnValue($connectionMock));
+        $resourceMock->expects($this->any())->method('getConnection')->willReturn($connectionMock);
         $selectMock = $this->getMockBuilder(\Magento\Framework\DB\Select::class)
             ->setMethods(['getPart', 'setPart', 'from', 'columns'])
             ->setConstructorArgs([$connectionMock, $selectRendererMock])
             ->getMock();
-        $connectionMock->expects($this->any())->method('select')->will($this->returnValue($selectMock));
+        $connectionMock->expects($this->any())->method('select')->willReturn($selectMock);
 
         $this->model = new \Magento\Indexer\Model\ResourceModel\Mview\View\State\Collection(
             $entityFactoryMock,
